Splitted funnel3d series into several files.
