+++ js/Series/Funnel3D/Funnel3DPoint.js
@@ -1,0 +1,63 @@
+/* *
+ *
+ *  Highcharts funnel3d series module
+ *
+ *  (c) 2010-2020 Highsoft AS
+ *
+ *  Author: Kacper Madej
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+'use strict';
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import BaseSeries from '../../Core/Series/Series.js';
+var ColumnSeries = BaseSeries.seriesTypes.column;
+import U from '../../Core/Utilities.js';
+var extend = U.extend;
+import '../Core/Renderer/SVG/SVGRenderer.js';
+/* *
+ *
+ *  Class
+ *
+ * */
+var Funnel3DPoint = /** @class */ (function (_super) {
+    __extends(Funnel3DPoint, _super);
+    function Funnel3DPoint() {
+        /* *
+         *
+         *  Properties
+         *
+         * */
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        _this.dlBoxRaw = void 0;
+        _this.options = void 0;
+        _this.series = void 0;
+        _this.y = void 0;
+        return _this;
+    }
+    return Funnel3DPoint;
+}(ColumnSeries.prototype.pointClass));
+extend(Funnel3DPoint.prototype, {
+    shapeType: 'funnel3d'
+});
+/* *
+ *
+ *  Default Export
+ *
+ * */
+export default Funnel3DPoint;
